[Hidden] reduce complexity by validating props at the top level api
