+++ src/Hidden/HiddenJs.js
@@ -1,14 +1,16 @@
 // @flow
-/**
- * Responsively hides children by omission.
- */
-import React, { Element, isValidElement } from 'react';
+import React, { Element } from 'react';
 import { keys as breakpoints } from '../styles/breakpoints';
 import withWidth, { isWidthDown, isWidthUp } from '../utils/withWidth';
-import type { HiddenProps } from './Hidden';
-import { defaultProps } from './Hidden';
+import type { HiddenProps } from './types';
 
 type Props = HiddenProps & {
+  /**
+   * If string or Function, component is used as the root node and all other props are passed
+   * including children.
+   * If an Element, it will be rendered as-is and no other props are propagated.
+   */
+    component: string | Function | Element<*>,
   /**
    * @ignore
    * width prop provided by withWidth decorator
@@ -16,10 +18,13 @@ type Props = HiddenProps & {
     width: string,
 };
 
+/**
+ * Responsively hides by omission.
+ */
 function HiddenJs(props: Props): ?Element<*> {
   const {
     children,
-    component,
+    component: ComponentProp,
     only,
     xsUp, // eslint-disable-line no-unused-vars
     smUp, // eslint-disable-line no-unused-vars
@@ -35,11 +40,9 @@ function HiddenJs(props: Props): ?Element<*> {
     ...other
   } = props;
 
-  // workaround: see https://github.com/facebook/flow/issues/1660#issuecomment-297775427
-  const ComponentProp = component || defaultProps.component;
   let visible = true;
 
-  // `only` takes priority.
+  // `only` check is faster to get out sooner if used.
   if (only) {
     if (Array.isArray(only)) {
       for (let i = 0; i < only.length; i += 1) {
@@ -75,19 +78,17 @@ function HiddenJs(props: Props): ?Element<*> {
     return null;
   }
 
-  // render any Element exactly as given
-  if (isValidElement(ComponentProp)) {
-    return React.Children.only(ComponentProp);
+  // render `string | Function` with any optional props
+  if (typeof ComponentProp === 'string' || typeof ComponentProp === 'function') {
+    return (
+      <ComponentProp {...other}>
+        {children}
+      </ComponentProp>
+    );
   }
 
-  // render `string | Function` with any optional props
-  return (
-    <ComponentProp {...other}>
-      {children}
-    </ComponentProp>
-  );
+  // render any Element exactly as given
+  return React.Children.only(ComponentProp);
 }
-
-HiddenJs.defaultProps = defaultProps;
 
 export default withWidth()(HiddenJs);
